+++ flattentool/tests/test_docs.py
@@ -136,7 +136,7 @@ def test_example_in_doc(root, filename):
 
 
 def test_expected_number_of_examples_in_docs_data():
-    expected = 66
+    expected = 67
     # See _get_examples_in_docs_data()
     if sys.version_info[:2] != (3, 8):
         expected -= 3