+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Generate templates that are correct for WKT <-> geojson conversion
+
 ## [0.21.0] - 2023-06-23
 
 ### Added